Improve performance of DiscoveryService.CalculateSourceLocationTrackingPositions
Profiling on a solution with a lot of Gherkin features shows that a significant amount of resources is taken in Deveroom when doing the repeated calls to `ResolveMefDependency` and `GetWpfTextViewFromFilePath` when calculating the tracking positions.

Here we reduce the number of calls to these methods:
- only call ResolveMefDependency once for each project being discovered
- group tracking positions by source file, so that we can call `GetWpfTextViewFromFilePath` once per source file only

After these changes, the resource usage for tracking position calculations goes down significantly.